Use an unsigned long when indexing into mfchashtbl[] and mf6ctable[].  This matches the types used when computing hash indices and the type of the maximum size of mfchashtbl[].
PR:		kern/181821
Submitted by:	Sven-Thorsten Dietrich <sven@vyatta.com> (IPv4)
MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@255248 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/netinet6/ip6_mroute.c
@@ -576,7 +576,7 @@ int
 X_ip6_mrouter_done(void)
 {
 	mifi_t mifi;
-	int i;
+	u_long i;
 	struct mf6c *rt;
 	struct rtdetq *rte;
 
@@ -1341,7 +1341,7 @@ expire_upcalls(void *unused)
 {
 	struct rtdetq *rte;
 	struct mf6c *mfc, **nptr;
-	int i;
+	u_long i;
 
 	MFC6_LOCK();
 	for (i = 0; i < MF6CTBLSIZ; i++) {
